To fault isolate I am no longer using AWS S3 locally. Uncommented the AWS S3 section in 'settings.py' temporarily.
CSS classes in HTML templates and 'base.css'.
Updated the (Shopping) Cart to use cards instead of a table as the table doesn't work well on mobiles. Added styling to the cards in `base.css`.

Added a dynamic `<h3>` header to the Shop for All Training, Courses, and Coaching checking request.path for the relevant link contents. This shows the user/buyer what products s/he is looking at.

Copied images to 'static/img' as this is where my images will be for both local and AWS S3 static images in future.

Documentation: updated 'README.md' with all of the above.