'settings.py' has local file environ loaded for the development.
Authenticated user can no longer access the register page, and is redirected to the 'dashboard'. Added validation and message to the register(), with a redirect. A different solution for the login page as I'm using Django Authentication that does not have a NOT @login_required or similar and I haven't extended it either. When an authenticated user accesses the login page via URL they are provided with an HTML message, and two links. One to access Logout or Dashboard.