Auto-fill of payment form. Authenticated users have the payment form auto-filled from the user and profile forms. Anonymous users receive and empty form that must be filled in before payment.
Updated the data model for orders to use post_code instead of postal_code, the same as for the profile to reuse the profile form for the payment form. Executed:
python3 manage.py makemigrations
python3 manage.py migrate

'orders/admin.py' renamed postal_code field to post_code to reuse for the payment form fields.

'orders/forms.py' renamed postal_code field to post_code to reuse for the payment form fields.

'orders/models.py' renamed postal_code field to post_code to reuse for the payment form fields.

'orders/views.py' updated the documentation string to explain order_create(). order_create() on GET will check if the user is authenticated, and if so create three forms, one empty and two instantiated on the user. If the user is anonymous (not authenticated) three forms are created empty. All three are used by the 'orders/templates/create.html' template to generate a filled in for authenticated users and an empty payment form for anonymous users. When the order is placed the filled in fields create the order form populating the order fields in the database.