'base.css': MaterializeCSS 1.0.0 overrides to provide a consistent look-and-feel. Adjusting the max-height and height to ensure the spacing of the cards is consistent as is the information in them. Card content is aligned to the bottom of the card, and this looks better: ----- .card.small .card-image+.card-content, .card.medium .card-image+.card-content, .card.large .card-image+.card-content {     max-height: 60%; }
.card.large {
    height: 600px;
}

.card-content {
    position: absolute;
    bottom: 0;
}
-----
'python3 manage.py collectstatic' -> AWS S3 Bucket.

'list.html': <div class="card medium"> medium-sized cards together with the CSS updates in layout and styling provide a consistent look-and-feel when shopping for courses and coaching.

'detail.html': <div class="card large"> large-sized cards containall the infomration on the card without overflow, and despite that some cards contain more information and larger images. Together with the CSS updates in layout and styling provide a consistent look-and-feel when shopping for courses and coaching.

Test:
- Tested the new card layouts across all device sizes locally.
- After GitHub update and to Heroku, further tests in production to follow.