Updated error page template '500.html', removed two links related to a previous image used (since removed).
'orders/models.py' added 'username' to the Order model for the updated Orders View on the Dashboard. Executed:
- 'python3 manage.py makemigrations'
- 'python3 manage.py migrate'

'orders/admin.py' added a the username field to the admin view.

'orders/views.py' moved 'form = OrderCreateForm()' out from the if-else statment for clarity. has no effect on performance. Added code to update the new username field with the current authenticated user or the anonymous user, to use this field (username) insetad of email address as a reference for the Orders View in the Dashboard.

'account/views.py' removed an errant print statement.